requirements.txt
cycler==0.10.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2
django-import-export==2.4.0
et-xmlfile==1.0.1
jdcal==1.4.1
kiwisolver==1.3.1
MarkupPy==1.14
matplotlib==3.3.3
numpy==1.19.3
odfpy==1.4.1
openpyxl==3.0.5
pandas==1.1.4
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
xlrd==1.2.0
xlwt==1.3.0
